+++ spec/scss_lint/linter/vendor_prefix_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe SCSSLint::Linter::VendorPrefixes do
+describe SCSSLint::Linter::VendorPrefix do
   context 'when no vendor-prefix is used' do
     let(:scss) { <<-SCSS }
       div {